refactor: split action_handler.go god file into crud / dispatch / params / schedule / validators (audit F005)
`internal/api/action_handler.go` was 1978 LOC / 39 functions and
mixed five distinct concerns into one file: per-RPC validators,
the proto-message extraction dispatch table, single-action CRUD
handlers, multi-target dispatch fan-outs, and the schedule
encode/decode helpers. Audit F005 has been open since the first
audit pass; this commit closes it.

# Split topology (same package, no API change)

- `action_handler.go` (270 LOC) — keeps the `ActionHandler` type,
  the `NewActionHandler` constructor, and the read-side translators
  (`actionToProto`, `executionToProto`, `statusToString`,
  `stringToStatus`, `loadLiveOutput`) that don't fit either the
  CRUD or the dispatch section cleanly. Header comment maps every
  removed symbol to its new home so `grep ActionHandler` still
  lands on the right file.
- `action_crud.go` (450 LOC) — single-action CRUD: CreateAction,
  GetAction, ListActions, RenameAction, UpdateActionDescription,
  UpdateActionParams, DeleteAction, plus the
  signature-lifecycle helpers `computeActionSignature` /
  `persistActionSignature` / `rollbackUnsignedCreate`.
- `action_dispatch.go` (820 LOC) — Dispatch* fan-outs and the
  in-flight execution reads: DispatchAction, DispatchToMultiple,
  DispatchAssignedActions, DispatchActionSet, DispatchDefinition,
  DispatchToGroup, DispatchInstantAction, GetExecution,
  ListExecutions, CancelExecution + isInstantActionType.
- `action_validators.go` (350 LOC) — validateCreateActionParams,
  validateUpdateActionParams, validateInlineAction,
  validateShellScriptChoice, validateAgentUpdateParams,
  actionParamsMatchType.
- `action_params.go` (155 LOC) — serializeProtoParams,
  extractCreateActionParamsMsg, extractUpdateActionParamsMsg,
  extractActionParamsMsg. Single point of truth for the
  action-type → proto.Message dispatch.
- `action_schedule.go` (60 LOC) — scheduleToMap / scheduleFromJSON.

# Conventions

- Every new file has a top-level Package-doc that names the
  audit ID and lists which symbols moved there from the original
  god file, so the next reader can navigate without `grep`.
- Imports trimmed per file: each uses only the packages it
  actually needs (no copy-paste of the original 14-import block).
- No behaviour change — purely a topology refactor; the public
  `ActionHandler` interface is identical.

Refs audit F005.
